# services/api-gateway/docs/openapi.json
{
    "openapi": "3.0.0",
    "info": {
        "description": "Microservices API Gateway with command and query services. Supports CQRS pattern with separate command and query endpoints.",
        "title": "KKM API Gateway",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@kkm-project.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "paths": {
...
        "/invoices": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Create a new invoice. Requires admin or manager role.",
                "tags": [
                    "Invoices"
                ],
                "summary": "Create invoice",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.Invoice"
                            }
                        }
                    },
                    "description": "Invoice data with all required fields",
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Invoice successfully created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/models.Invoice"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input - validation errors",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Forbidden - requires admin or manager role",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/invoices/{id}": {
            "put": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Update an existing invoice",
                "tags": [
                    "Invoices"
                ],
                "summary": "Update invoice",
                "parameters": [
                    {
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/models.Invoice"
                            }
                        }
                    },
                    "description": "Updated invoice data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/models.Invoice"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{id}/accept-reject": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Accept or reject an existing invoice",
                "tags": [
                    "Invoices"
                ],
                "summary": "Accept or reject invoice",
                "parameters": [
                    {
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "accept": {
                                        "type": "boolean"
                                    },
                                    "reason": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "description": "Accept/reject decision",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/models.Invoice"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{id}/revoke": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Revoke an existing invoice",
                "tags": [
                    "Invoices"
                ],
                "summary": "Revoke invoice",
                "parameters": [
                    {
                        "description": "Invoice ID",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "reason": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "description": "Revocation reason"
                },
                "responses": {
                    "200": {
                        "description": "OK",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/models.Invoice"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/invoices/{id}/sign": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Sign an existing invoice with digital signature",
                "tags": [
                    "Invoices"
                ],
                "summary": "Sign invoice",
                "parameters": [
                    {
                        "description": "Invoice ID (UUID)",
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "signature_data": {
                                        "type": "string"
                                    }
                                }
                            }
                        }
                    },
                    "description": "Signature data",
                    "required": true
                },
                "responses": {
                    "200": {
                        "description": "Invoice signed successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "allOf": [
                                        {
                                            "$ref": "#/components/schemas/models.APIResponse"
                                        },
                                        {
                                            "type": "object",
                                            "properties": {
                                                "data": {
                                                    "$ref": "#/components/schemas/models.Invoice"
                                                }
                                            }
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Invalid input",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Internal server error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/models.APIResponse"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
    "servers": [
        {
            "url": "http://localhost:8080/api/v1"
        },
        {
            "url": "https://localhost:8080/api/v1"
        }
    ],
    "components": {
...
        "securitySchemes": {
            "BearerAuth": {
                "description": "Type \"Bearer\" followed by a space and JWT token.",
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        },
        "schemas": {
...
            "models.APIError": {
                "type": "object",
                "properties": {
                    "code": {
                        "type": "string"
                    },
                    "details": {
                        "type": "string"
                    },
                    "message": {
                        "type": "string"
                    }
                }
            },
            "models.APIResponse": {
                "type": "object",
                "properties": {
                    "data": {},
                    "error": {
                        "$ref": "#/components/schemas/models.APIError"
                    },
                    "meta": {
                        "$ref": "#/components/schemas/models.MetaData"
                    },
                    "success": {
                        "type": "boolean"
                    }
                }
            },
...
            "models.Invoice": {
                "type": "object",
                "properties": {
                    "created_at": {
                        "type": "integer"
                    },
                    "delivery_date": {
                        "type": "string"
                    },
                    "id": {
                        "type": "string"
                    },
                    "invoice_date": {
                        "type": "string"
                    },
                    "invoice_number": {
                        "type": "string"
                    },
                    "is_resident": {
                        "type": "boolean"
                    },
                    "note": {
                        "type": "string"
                    },
                    "status": {
                        "type": "string"
                    },
                    "total_amount": {
                        "type": "number"
                    },
                    "updated_at": {
                        "type": "integer"
                    }
                }
            },
...
            "models.MetaData": {
                "type": "object",
                "properties": {
                    "page": {
                        "type": "integer"
                    },
                    "page_size": {
                        "type": "integer"
                    },
                    "total_count": {
                        "type": "integer"
                    },
                    "total_pages": {
                        "type": "integer"
                    }
                }
            },
...
        }
    }
}